+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [Unreleased]
 
+## Fixed
+
+- call lexer after reduction in LR parser as the set of possible tokens may change.
+
 
 # [0.5.0] - 2024-01-23